options.py

+++ options.py
@@ -34,7 +34,7 @@ if vol_choice == "Historical":
     rolling_std = returns.rolling(window=30).std()
     sigma_last = rolling_std.iloc[-1]
     sigma = sigma_last * np.sqrt(252)  # last value
-    sigma_display = float(round((sigma), 4))
+    sigma_display = round((sigma), 4)
     st.write(f"Historical Volatility (σ) calculated from past 30 days: ", {sigma_display*100:.2f}%)
 else:
         # Let user enter custom volatility via slider
